+++ src/styles/sidebar/ArchiveButtonStyle.js
@@ -17,6 +17,25 @@ export const ArchiveList = styled.div`
   /* 스크롤 */
   max-height: ${({ isScrollable }) => (isScrollable ? '300px' : 'auto')};
   overflow-y: ${({ isScrollable }) => (isScrollable ? 'auto' : 'hidden')};
+
+  /* 스크롤바 스타일 추가 */
+  &::-webkit-scrollbar {
+    width: 6px;
+  }
+
+  &::-webkit-scrollbar-track {
+    background: transparent;
+    margin: 5px;
+  }
+
+  &::-webkit-scrollbar-thumb {
+    background: #e0e0e0;
+    border-radius: 50px;
+  }
+
+  &::-webkit-scrollbar-thumb:hover {
+    background: #bdbdbd;
+  }
 `;
 
 /* 클릭 시 회색 배경 */
@@ -43,7 +62,7 @@ export const ListBox = styled.div`
   display: flex;
   align-items: center;
   justify-content: space-between;
-  padding: 15px 3px 15px 1px;
+  padding: 15px 1px;
   margin: 0px;
   cursor: pointer;
   border-bottom: 1px solid #ececec;
