+++ src/styles/sidebar/CoverSelectionStyle.js
