+++ src/styles/sidebar/NameInputStyle.js
