+++ src/styles/sidebar/PublicSelectionStyle.js
